clore latest/clore-image-resizer-api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.0.1
torch==2.5.0
torchvision==0.20.0
git+https://github.com/huggingface/diffusers.git
transformers>=4.38.0
tokenizers>=0.18.0
safetensors==0.4.1
accelerate>=0.32.0
peft>=0.4.0
hf-transfer>=0.1.3
numpy<2
pydantic==2.5.0
scipy>=1.11.0
huggingface-hub>=0.19.0
protobuf>=3.20.0
sentencepiece>=0.1.99
azure-storage-queue>=12.0.0
azure-storage-blob>=12.0.0
python-dotenv>=0.19.0
openai>=1.0.0
psycopg2-binary>=2.9.0
aiofiles>=23.1.0
websockets>=11.0
requests>=2.31.0
httpx>=0.25.0
